Rename PeerSessions->MultipeerPeerSessions
