+++ plugin.xml
@@ -63,12 +63,12 @@
         <source-file src="src/ios/THEMultipeerClient.m" />
         <header-file src="src/ios/THEMultipeerServer.h" />
         <source-file src="src/ios/THEMultipeerServer.m" />
-        <header-file src="src/ios/THEPeerSession.h" />
-        <source-file src="src/ios/THEPeerSession.m" />
-        <header-file src="src/ios/THEPeerClientSession.h" />
-        <source-file src="src/ios/THEPeerClientSession.m" />
-        <header-file src="src/ios/THEPeerServerSession.h" />
-        <source-file src="src/ios/THEPeerServerSession.m" />
+        <header-file src="src/ios/THEMultipeerPeerSession.h" />
+        <source-file src="src/ios/THEMultipeerPeerSession.m" />
+        <header-file src="src/ios/THEMultipeerClientSession.h" />
+        <source-file src="src/ios/THEMultipeerClientSession.m" />
+        <header-file src="src/ios/THEMultipeerServerSession.h" />
+        <source-file src="src/ios/THEMultipeerServerSession.m" />
         <header-file src="src/ios/THENetworkingRelay.h" />
         <source-file src="src/ios/THENetworkingRelay.m" />
         <header-file src="src/ios/THENetworkingClientRelay.h" />
